Refactor connection management into shared core
