+++ InteractiveClassroom/View/Client/Teacher/TeacherDashboardView.swift
@@ -86,10 +86,9 @@ struct TeacherDashboardView: View {
     }
 }
 #Preview {
-    let manager = PeerConnectionManager()
-    manager.students = ["Alice", "Bob"]
-    let courseService = CourseSessionService(manager: manager)
-    let pairing = PairingService(manager: manager)
+    let pairing = PairingService()
+    pairing.students = ["Alice", "Bob"]
+    let courseService = CourseSessionService(manager: pairing)
     return NavigationStack { TeacherDashboardView() }
         .environmentObject(courseService)
         .environmentObject(pairing)
